+++ internal/elasticsearch/index/data_stream_lifecycle/update.go
@@ -1,0 +1,11 @@
+package data_stream_lifecycle
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func (r *Resource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	resp.Diagnostics.Append(r.create(ctx, req.Plan, &resp.State)...)
+}